Refactored entry table reading and writing.
